Create 1.TwoSum.py
Here we use a map to store list element and its indices:
We iterate over the list, and for current element we check if target-current value exists in map. Below are the 2 possibilities:
If it does not, we insert the element x as key into map with its list index as value and move on to next element in the list.
If x exists in the map then we return the indices of current element and the index(value) of x(key) from the map.